First Pass at replacing OpenTK on Desktop.
New files

OpenAL.cs
OpenGL.cs
OpenGL.Common.cs
OpenGL.SDL.cs

NOTE : The attributes on the delegate declarations might look
weird but they are REQUIRED for iOS. Since the end goal is to use
exactly the same OpenGL.cs file for all the GL platforms we need
these. The idea being that platforms line SDL and Android that do
not need [MonoNativeFunctionWrapper] will just use

	OpenGL.Common.cs

which defines a place holder.

When we move to mobile as well new files such as

	OpenGL.Android.cs
	OpenGL.iOS.cs

will be added to implement the partial

	LoadPlatformEntryPoints

and

	EntryPointHelper

EntryPointHelper just provide the implementation for the
GetProcessAddress funciton. For SDL we MUST use

	SDL_GL_GetProcAddress

as well as provide the various GL attributes via

	SDL_GL_Set_Attribute